chore: rename function name and move to adyen checkout handler
