--- src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/validateComponents.js
@@ -1,18 +1,0 @@
-const store = require('../../../../store');
-
-module.exports.validateComponents = function validateComponents() {
-  const customMethods = {};
-
-  if (store.selectedMethod in customMethods) {
-    customMethods[store.selectedMethod]();
-  }
-
-  document.querySelector('#adyenStateData').value = JSON.stringify(
-    store.stateData,
-  );
-  if (store.partialPaymentsOrderObj) {
-    document.querySelector('#adyenPartialPaymentsOrder').value = JSON.stringify(
-      store.partialPaymentsOrderObj,
-    );
-  }
-};
