+++ src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/checkout_services/adyenCheckoutServices.js
@@ -3,7 +3,7 @@ const collections = require('*/cartridge/scripts/util/collections');
 
 function processPayment(order, handlePaymentResult, req, res, emit) {
   res.json({
-    error: true,
+    error: false,
     adyenAction: handlePaymentResult.action,
     orderID: order.orderNo,
     orderToken: order.orderToken,
